+++ src/app/client/page.tsx
@@ -1,12 +1,16 @@
 import React from "react";
 import { ClientList } from "./client-list";
-import { ClientSet } from "./client-set";
+import { ClientForm } from "./client-form";
 
 export default function ClientsPage() {
   return (
     <React.Fragment>
+      <br />
+      <ClientForm />
+      <br />
+      <hr />
+      <br />
       <ClientList />
-      <ClientSet />
     </React.Fragment>
   );
 }