+++ src/app/event/page.tsx
@@ -1,0 +1,14 @@
+import { EventForm } from "./event-form";
+import { EventList } from "./event-list";
+
+function Events() {
+  return (
+    <main>
+      <EventForm />
+      <br />
+      <EventList />
+    </main>
+  );
+}
+
+export default Events;