+++ src/lib/features/event/event.ts
@@ -1,0 +1,11 @@
+export interface Event {
+  clientId: string;
+  name: string;
+  eventId?: string;
+  locationId?: string;
+  organizerId?: string;
+  description?: string;
+  category?: string;
+  createdOn?: string;
+  lastUpdatedOn?: string;
+}