+++ src/lib/features/scheduled-event/scheduled-event-day.ts
@@ -1,0 +1,9 @@
+export enum ScheduledEventDay {
+  MON,
+  TUE,
+  WED,
+  THU,
+  FRI,
+  SAT,
+  SUN,
+}