+++ src/lib/features/scheduled-event/scheduled-event-interval.ts
@@ -1,0 +1,6 @@
+export enum ScheduledEventInterval {
+  DAILY,
+  WEEKLY,
+  WEEKDAYS,
+  WEEKENDS,
+}